+++ src/components/SideBar.jsx
@@ -1,71 +1,140 @@
 import React from "react";
 import { styled } from "styled-components";
+import { useState } from "react";
+import { categoryOptions, subcategoryOptions } from "./MyPost";
 
-const Sidebar = styled.div`
-  margin-top: 180px;
-  margin-left: 0%;
-  width: 250px;
-  border: none;
-  border-radius: 20px;
-
-  text-align: center;
+const AllList = styled.button`
+  margin-top: 200px;
+  /* height: 540px; */
+  width: 220px;
   padding: 10px;
-
-  background-color: #f5f5f5;
-  padding-bottom: 10%;
-  box-shadow: 2px 2px 4px rgba(0, 0, 0, 0.3);
-`;
-
-const Sidebar2 = styled.div`
-  margin-top: 2%;
-  margin-left: 0%;
-  width: 250px;
-  border: none;
   border-radius: 20px;
-
-  text-align: center;
-  padding: 10px;
-
-  background-color: #f5f5f5;
-  padding-bottom: 10%;
-  box-shadow: 2px 2px 4px rgba(0, 0, 0, 0.3);
+  display: flex;
+  flex-direction: column;
+  justify-content: flex-start;
+  align-items: center;
+  /* margin-bottom: 10px; */
+  padding-top: 20px;
+  border: none;
+  position: fixed;
 `;
-
-const SideBtn = styled.button`
+const CategoryFont = styled.div`
+  margin-bottom: 20px;
+  font-size: 20px;
+  font-weight: bold;
+`;
+const List = styled.button`
+  display: flex;
+  /* align-items: center; */
+  justify-content: center;
+  flex-direction: column;
   background-color: transparent;
   border: none;
-  width: 100px;
-  height: 40px;
-  margin: 10px auto 5px auto;
-  &:hover {
-    transform: scale(1.2);
-  }
-  cursor: pointer;
-
-  font-size: 20px;
-  display: flex;
+  /* font-weight: bold; */
+  font-size: 18px;
+  padding: 15px;
+  width: 180px;
+  /* margin-top: 10px; */
+`;
+const SmallLists = styled.div`
+  display: ${(props) => (props.isOpen ? "flex" : "none")};
+  flex-direction: column;
+  /* align-items: center; */
+  justify-content: center;
+`;
+const SmallList = styled.button`
+  font-size: 15px;
+  padding-top: 10px;
+  text-align: left;
+  border: none;
 `;
 
-function SideBar() {
+function SideBar2() {
+  console.log(subcategoryOptions);
+  const initialallLists = [
+    {
+      id: 1,
+      list: "💰 경제",
+      sublist: subcategoryOptions["경제"].map((option) => option.value),
+      // sublist: ["📈 주식", "💸 가상화폐"],
+      isOpen: false,
+    },
+    {
+      id: 2,
+      list: "🐶 애완동.식물",
+      sublist: subcategoryOptions["애완동식물"].map((option) => option.value),
+      // sublist: ["🍯 꿀팁", "💳 쇼핑"],
+      isOpen: false,
+    },
+    {
+      id: 3,
+      list: "🚙 여행",
+      sublist: subcategoryOptions["여행"].map((option) => option.value),
+      // sublist: ["🚅 국내", "🛩️ 해외"],
+      isOpen: false,
+    },
+    {
+      id: 4,
+      list: "🎧 음악",
+      sublist: subcategoryOptions["음악"].map((option) => option.value),
+      // sublist: ["🎤 추천", "🎹 정보"],
+      isOpen: false,
+    },
+    {
+      id: 5,
+      list: "🍀 기타",
+      sublist: subcategoryOptions["기타"].map((option) => option.value),
+      // sublist: ["🎤 추천", "🎹 정보"],
+      isOpen: false,
+    },
+  ];
+
+  // const [isOpen, setIsOpen] = useState(false);
+  // const handleList = () => {
+  //   setIsOpen(!isOpen);
+  // };
+  const handleList = (id) => {
+    const updatedLists = allLists.map((item) => {
+      if (item.id === id) {
+        return {
+          ...item,
+          isOpen: !item.isOpen,
+        };
+      }
+      return item;
+    });
+    setAllLists(updatedLists);
+  };
+
+  const [allLists, setAllLists] = useState(initialallLists);
+
   return (
-    <div>
-      <Sidebar>
-        <h3>▼ Category</h3>
-        <br />
-        <SideBtn>여행</SideBtn>
-        <br />
-        <SideBtn>카테고리2</SideBtn>
-        <br />
-        <SideBtn>카테고리3</SideBtn>
-      </Sidebar>
-      <div>
-        <Sidebar2>
-          <h3>경제 API</h3>
-          <br />
-        </Sidebar2>
-      </div>
-    </div>
+    <AllList>
+      <CategoryFont>♞ Category</CategoryFont>
+      {allLists.map((allList) => {
+        return (
+          <List
+            className="큰목차"
+            key={allList.id}
+            onClick={() => handleList(allList.id)}
+          >
+            <div>{allList.list}</div>
+            <SmallLists className="작은목차" isOpen={allList.isOpen}>
+              {allLists.map((subList, i) => {
+                if (allList.sublist[i]) {
+                  return (
+                    allList.sublist.length > 0 && (
+                      <SmallList key={i}>{allList.sublist[i]}</SmallList>
+                    )
+                  );
+                }
+                return null;
+              })}
+            </SmallLists>
+          </List>
+        );
+      })}
+    </AllList>
   );
 }
-
-export default SideBar;
+export default SideBar2;
