+++ src/pages/Main.jsx
@@ -1,50 +1,10 @@
 import React from "react";
 import TopBar from "../components/TopBar";
-import SideBar2 from "../components/SideBar2";
+import SideBar from "../components/SideBar";
 import Contents from "../components/Contents";
 import Post from "../components/Post";
-// import { useState, useEffect } from "react";
-import { getDocs, collection } from "firebase/firestore";
-import { db } from "../service/firebase";
 
 function App() {
-  const [data, setData] = useState([]);
-  const [isLoading, setIsLoading] = useState(false);
-
-  const fetchData = async () => {
-    try {
-      setIsLoading(true);
-      const querySnapshot = await getDocs(collection(db, "posts"));
-      const fetchedData = querySnapshot.docs.map((doc) => doc.data());
-      setData((prevData) => [...prevData, ...fetchedData]);
-      setIsLoading(false);
-    } catch (error) {
-      console.error("Error fetching data:", error);
-      setIsLoading(false);
-    }
-  };
-
-  useEffect(() => {
-    const handleScroll = () => {
-      if (
-        window.innerHeight + window.scrollY >= document.body.offsetHeight &&
-        !isLoading
-      ) {
-        fetchData();
-      }
-    };
-
-    window.addEventListener("scroll", handleScroll);
-    return () => {
-      window.removeEventListener("scroll", handleScroll);
-    };
-  }, [isLoading]); // isLoading을 의존성으로 포함
-
-  useEffect(() => {
-    fetchData();
-  }, []);
-
-  //---------------------------
   return (
     <>
       <div
@@ -55,16 +15,11 @@ function App() {
       >
         <div style={{ display: "flex" }}>
           <TopBar />
-          <SideBar2 />
+          <SideBar />
           <Contents />
           <Post />
         </div>
       </div>
-
-      {/* 데이터 출력 */}
-      {data.map((item, index) => (
-        <div key={index}>{item.title}</div>
-      ))}
     </>
   );
 }
